Split citizen_info.rs (614 lines) into citizen_info/ subdirectory
Split crates/ui/src/citizen_info.rs into focused modules under citizen_info/:
- mod.rs: thin facade with re-exports
- resources.rs: SelectedCitizen and FollowCitizen resources
- names.rs: deterministic name generation from Entity index + Gender
- display.rs: label, color, and UI bar helpers
- systems.rs: selection detection, camera follow, info panel rendering
- plugin.rs: CitizenInfoPlugin registration
- tests.rs: unit tests for helpers and resources

All public items re-exported from mod.rs — no external import changes needed.
