+++ crates/ui/src/citizen_info/names.rs
@@ -1,0 +1,91 @@
+//! Deterministic citizen name generation from Entity index and Gender.
+
+use bevy::prelude::*;
+use simulation::citizen::Gender;
+
+const FIRST_NAMES_M: &[&str] = &[
+    "James", "John", "Robert", "Michael", "David", "William", "Richard", "Joseph", "Thomas",
+    "Daniel", "Matthew", "Anthony", "Mark", "Steven", "Paul", "Andrew", "Joshua", "Kenneth",
+    "Kevin", "Brian", "George", "Timothy", "Ronald", "Edward", "Jason", "Jeffrey", "Ryan", "Jacob",
+    "Gary", "Nicholas", "Eric", "Jonathan",
+];
+
+const FIRST_NAMES_F: &[&str] = &[
+    "Mary",
+    "Patricia",
+    "Jennifer",
+    "Linda",
+    "Barbara",
+    "Elizabeth",
+    "Susan",
+    "Jessica",
+    "Sarah",
+    "Karen",
+    "Lisa",
+    "Nancy",
+    "Betty",
+    "Margaret",
+    "Sandra",
+    "Ashley",
+    "Emily",
+    "Donna",
+    "Michelle",
+    "Carol",
+    "Amanda",
+    "Dorothy",
+    "Melissa",
+    "Deborah",
+    "Stephanie",
+    "Rebecca",
+    "Sharon",
+    "Laura",
+    "Cynthia",
+    "Kathleen",
+    "Amy",
+    "Angela",
+];
+
+const LAST_NAMES: &[&str] = &[
+    "Smith",
+    "Johnson",
+    "Williams",
+    "Brown",
+    "Jones",
+    "Garcia",
+    "Miller",
+    "Davis",
+    "Rodriguez",
+    "Martinez",
+    "Hernandez",
+    "Lopez",
+    "Wilson",
+    "Anderson",
+    "Thomas",
+    "Taylor",
+    "Moore",
+    "Jackson",
+    "Martin",
+    "Lee",
+    "Thompson",
+    "White",
+    "Harris",
+    "Clark",
+    "Lewis",
+    "Robinson",
+    "Walker",
+    "Young",
+    "Allen",
+    "King",
+    "Wright",
+    "Hill",
+];
+
+pub fn citizen_name(entity: Entity, gender: Gender) -> String {
+    let idx = entity.index() as usize;
+    let first = match gender {
+        Gender::Male => FIRST_NAMES_M[idx % FIRST_NAMES_M.len()],
+        Gender::Female => FIRST_NAMES_F[idx % FIRST_NAMES_F.len()],
+    };
+    let last = LAST_NAMES[(idx / 31) % LAST_NAMES.len()];
+    format!("{} {}", first, last)
+}
